[refactor] Split 유틸을 StringUtils 로 합침
- Split은 결국 문자열을 콤마 기준으로 분리하는 역할이므로 StringUtils에
포함시켰다.
