--- src/main/java/racingcar/utils/Split.java
@@ -1,13 +1,0 @@
-package racingcar.utils;
-
-public class Split {
-
-    private static final String COMMA = ",";
-
-    private Split() {
-    }
-
-    public static String[] splitInputName(String values) {
-        return values.split(Split.COMMA);
-    }
-}
