+++ src/main/java/racingcar/utils/StringUtils.java
@@ -2,10 +2,15 @@ package racingcar.utils;
 
 public class StringUtils {
 
+    private static final String COMMA = ",";
     private static final String BLANK = " ";
     private static final String NO_BLANK = "";
 
     private StringUtils() {
+    }
+
+    public static String[] splitInputName(String values) {
+        return values.split(COMMA);
     }
 
     public static String removeAllBlank(String input) {
