--- src/test/java/racingcar/utils/SplitTest.java
@@ -1,19 +1,0 @@
-package racingcar.utils;
-
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-class SplitTest {
-    @Test
-    @DisplayName("자동차 이름을 쉼표(,) 기준으로 분리 테스트")
-    void splitNameTest() {
-        // given
-        String values = "pobi,crong,honux";
-        String[] expected = {"pobi", "crong", "honux"};
-        // when
-        String[] result = Split.splitInputName(values);
-        // then
-        Assertions.assertThat(result).containsExactly(expected);
-    }
-}
