+++ src/test/java/racingcar/utils/StringUtilsTest.java
@@ -1,10 +1,23 @@
 package racingcar.utils;
 
 import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
 class StringUtilsTest {
+
+    @Test
+    @DisplayName("자동차 이름을 쉼표(,) 기준으로 분리 테스트")
+    void splitNameTest() {
+        String values = "pobi,crong,honux";
+        String[] expected = {"pobi", "crong", "honux"};
+
+        String[] result = StringUtils.splitInputName(values);
+
+        Assertions.assertThat(result).containsExactly(expected);
+    }
 
     @ParameterizedTest
     @ValueSource(strings = {"po bi", "  pobi  "})
